+++ .github/CHANGELOG.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* `*messageContext.SendMessageString` and `*messageContext.SendMessageEmbed` 
 
 ## [0.1.1] - 2021-03-08
 ### Added
 * `kit.AllowBots`
 * `kit.DefaultAllowedMentions`
 ### Changed
-* Bow now ignores own messages and reactions
+* Bot now ignores own messages and reactions
 
 ## [0.1.0] - 2021-03-08
 ### Added